+++ packages/terra-date-picker/tests/nightwatch/DatePickerTests.jsx
@@ -13,6 +13,7 @@ const DatePickerTests = () => (
       <li><Link to="/tests/date-picker-tests/min-max">DatePicker - Min Max</Link></li>
       <li><Link to="/tests/date-picker-tests/on-change">DatePicker - On Change</Link></li>
       <li><Link to="/tests/date-picker-tests/start-date">DatePicker - Start Date</Link></li>
+      <li><Link to="/tests/date-picker-tests/inside-modal">DatePicker - Presented Inside Modal</Link></li>
     </ul>
   </div>
 );