+++ packages/terra-modal-manager/tests/nightwatch/modal-manager-spec.js
@@ -23,7 +23,7 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
   },
 
   'Ensures goBack is not provided to a single modal component': (browser) => {
@@ -31,9 +31,9 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1 button.go-back').to.not.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1 button.go-back').to.not.be.present;
   },
 
   'Closes the modal when dismiss is selected within modal with one modal component': (browser) => {
@@ -41,11 +41,11 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
 
     browser.click('#DemoContainer-1 .dismiss');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.not.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.not.be.present;
   },
 
   'Closes the modal when closeDisclosure is selected within modal': (browser) => {
@@ -53,11 +53,11 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
 
     browser.click('#DemoContainer-1 .close-disclosure');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.not.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.not.be.present;
   },
 
   'Maximizes/minimizes the modal when selected within modal': (browser) => {
@@ -65,18 +65,18 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
     browser.expect.element('#DemoContainer-1 .maximize').to.be.present;
 
     browser.click('#DemoContainer-1 .maximize');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
     browser.expect.element('#DemoContainer-1 .maximize').to.not.be.present;
     browser.expect.element('#DemoContainer-1 .minimize').to.be.present;
 
     browser.click('#DemoContainer-1 .minimize');
 
-    browser.expect.element('.terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
     browser.expect.element('#DemoContainer-1 .maximize').to.be.present;
     browser.expect.element('#DemoContainer-1 .minimize').to.not.be.present;
   },
@@ -86,22 +86,43 @@ module.exports = {
 
     browser.click('#root-component .disclose');
 
-    browser.expect.element('.terra-ModalManager-modal .terra-SlideGroup #DemoContainer-1').to.be.present;
+    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
 
     browser.click('#DemoContainer-1 .disclose');
 
     // Waiting here to ensure new component is presented and back button is clickable
-    browser.waitForElementPresent('.terra-Slide:not(.terra-Slide-enter-active):nth-child(2)', 1000);
+    browser.waitForElementPresent('[class*="slide-group"] > span > div[class*="slide"]:not([class*="enter-active"]):nth-child(2)', 1000);
 
-    browser.expect.element('.terra-ModalManager-modal .terra-SlideGroup #DemoContainer-1').to.be.present;
-    browser.expect.element('.terra-ModalManager-modal .terra-SlideGroup #DemoContainer-2').to.be.present;
-    browser.expect.element('.terra-SlideGroup #DemoContainer-2 button.go-back').to.be.present;
+    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-2').to.be.present;
+    browser.expect.element('[class*="slide-group"] #DemoContainer-2 button.go-back').to.be.present;
 
     browser.click('#DemoContainer-2 button.go-back');
 
     browser.waitForElementNotPresent('#DemoContainer-2', 1000);
 
-    browser.expect.element('.terra-ModalManager-modal .terra-SlideGroup #DemoContainer-1').to.be.present;
-    browser.expect.element('.terra-ModalManager-modal .terra-SlideGroup #DemoContainer-2').to.not.be.present;
+    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('.terra-ModalManager-modal [class*="slide-group"] #DemoContainer-2').to.not.be.present;
+  },
+
+  'Releases/Requests focus when selected within modal': (browser) => {
+    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/modal-manager-tests/demo`);
+
+    browser.click('#root-component .disclose');
+
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('#DemoContainer-1 .requestFocus').to.be.present;
+
+    browser.click('#DemoContainer-1 .requestFocus');
+
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('#DemoContainer-1 .requestFocus').to.not.be.present;
+    browser.expect.element('#DemoContainer-1 .releaseFocus').to.be.present;
+
+    browser.click('#DemoContainer-1 .releaseFocus');
+
+    browser.expect.element('[class*="slide-group"] #DemoContainer-1').to.be.present;
+    browser.expect.element('#DemoContainer-1 .requestFocus').to.be.present;
+    browser.expect.element('#DemoContainer-1 .releaseFocus').to.not.be.present;
   },
 };