+++ packages/terra-modal-manager/CHANGELOG.md
@@ -3,11 +3,14 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added gainFocus and loseFocus actions to support trapping and releasing focus in the modal manager.
 
 1.2.0 - (July 26, 2017)
 ------------------
 ### Changed
 * Updated spacing around react-docgen comments
+* Update nightwatch tests for Slide Group conversion to CSS modules
 
 1.1.0 - (July 18, 2017)
 ------------------