+++ packages/terra-site/CHANGELOG.md
@@ -3,11 +3,15 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added new example to display date picker and popup inside modal.
 
 1.2.0 - (July 26, 2017)
 ------------------
 ### Changed
+* Converted component to use CSS modules
 * Updated Search Field example to demonstrate onInvalidSearch.
+* Add Model Manager Test Link
 
 1.1.0 - (July 18, 2017)
 ------------------